+++ 1_data/raw-spice_filtered-z-score/state.json
@@ -1,0 +1,25 @@
+{
+  "_data_files": [
+    {
+      "filename": "data-00000-of-00005.arrow"
+    },
+    {
+      "filename": "data-00001-of-00005.arrow"
+    },
+    {
+      "filename": "data-00002-of-00005.arrow"
+    },
+    {
+      "filename": "data-00003-of-00005.arrow"
+    },
+    {
+      "filename": "data-00004-of-00005.arrow"
+    }
+  ],
+  "_fingerprint": "ef152e820cd35d2f",
+  "_format_columns": null,
+  "_format_kwargs": {},
+  "_format_type": "torch",
+  "_output_all_columns": false,
+  "_split": null
+}
